requirementts.txt
africastalking==1.2.4
certifi==2022.6.15
charset-normalizer==2.1.1
contextlib2==21.6.0
idna==3.3
numpy==1.23.2
pandas==1.4.4
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
schema==0.7.5
six==1.16.0
urllib3==1.26.12
